Updated last set of tests from mercury specific tests thus completing the refactor
