+++ .eslintignore
@@ -1,3 +1,3 @@
 node_modules
-lib
+/lib
 app.js
